Simplify legacy smart field import
